+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 Version counting is based on semantic versioning (Major.Feature.Patch)
 
 ## WIP (9.15.1)
+### YACReader
+* Don't use scroll animations on macos by default, it where hdpi scroll is most likely to be used.
 
 ### YACReaderLibrary
 * Improve flexibility of the open comic in third party app setting so more complex commands can be used. e.g. `open -a "/Applications/My Reader.app" "{comic_file_path}"`.